+++ src/ExtPack.h
@@ -86,7 +86,7 @@ ext_pack_error_t _send_to_ExtPack(unit_t unit, uint8_t data);
  * @param send_byte_delay_us The delay between sending two bytes in us.
  * @return EXT_PACK_SUCCESS on success, EXT_PACK_FAILURE if the function was aborted when sending an uint8_t because of an error while sending.
  */
-ext_pack_error_t send_String_to_ExtPack(unit_t unit, const uint8_t* data, uint8_t send_byte_delay_us);
+ext_pack_error_t send_String_to_ExtPack(unit_t unit, const uint8_t* data, uint16_t send_byte_delay_us);
 
 /**
  * Returns the duration a UART send operation to ExtPack needs to perform in the worst case in us.
