+++ src/ExtPack_U_SPI.c
@@ -15,7 +15,7 @@ ext_pack_error_t send_ExtPack_SPI_data_to_slave(unit_t unit, uint8_t slave_id, u
     return send_ExtPack_SPI_data(unit, data);
 }
 
-ext_pack_error_t send_ExtPack_SPI_String_to_slave(unit_t unit, uint8_t slave_id, const uint8_t* data, uint8_t send_byte_delay_us) {
+ext_pack_error_t send_ExtPack_SPI_String_to_slave(unit_t unit, uint8_t slave_id, const uint8_t* data, uint16_t send_byte_delay_us) {
     if(set_ExtPack_SPI_slave(unit, slave_id) == EXT_PACK_FAILURE) {
         return EXT_PACK_FAILURE;
     }
