+++ src/ExtPack_U_SPI.h
@@ -60,7 +60,7 @@ ext_pack_error_t send_ExtPack_SPI_data_to_slave(unit_t unit, uint8_t slave_id, u
  * @param send_byte_delay_us The delay between sending two bytes in us.
  * @return EXT_PACK_SUCCESS on success, EXT_PACK_FAILURE if the function was aborted when sending a char because of an error while sending.
  */
-ext_pack_error_t send_ExtPack_SPI_String_to_slave(unit_t unit, uint8_t slave_id, const uint8_t* data, uint8_t send_byte_delay_us);
+ext_pack_error_t send_ExtPack_SPI_String_to_slave(unit_t unit, uint8_t slave_id, const uint8_t* data, uint16_t send_byte_delay_us);
 
 /**
  * @brief Retrieves the last set SPI slave id of the given ExtPack SPI unit.
@@ -79,7 +79,7 @@ uint8_t get_ExtPack_data_SPI_current_slave(unit_t unit);
  * @param send_byte_delay_us The delay between sending two bytes in us.
  * @return EXT_PACK_SUCCESS on success, EXT_PACK_FAILURE if the function was aborted when sending an uint8_t because of an error while sending.
  */
-static inline ext_pack_error_t send_ExtPack_SPI_String(unit_t unit, const uint8_t* data, uint8_t send_byte_delay_us) {
+static inline ext_pack_error_t send_ExtPack_SPI_String(unit_t unit, const uint8_t* data, uint16_t send_byte_delay_us) {
     return send_String_to_ExtPack(_set_ExtPack_access_mode(unit, 00), data, send_byte_delay_us);
 }
